Finish client command line tool
